AG-1660: rename file
